--- Backlight.cpp
@@ -1,5 +1,0 @@
-#include "Backlight.h"
-
-void Backlight::initBacklight() {
-
-}
