--- Backlight.h
@@ -1,4 +1,0 @@
-class Backlight{
-  public:
-  void initBacklight();
-};
